+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project are documented here. The format is human-written
 and organized by user-visible concern.
 
-## Unreleased
+## 0.1.4 - 2026-08-17
 
 ### Fixed